Refactor core/cluster separation
